Update setupLetsEncrypt.php Deb11/12
Quick fix, not quite proper .... But works for now until let's encrypt changes something.

Real solution would be a let's encrypt docker since it's python.

+++ scripts/util/setupLetsEncrypt.php
@@ -3,6 +3,7 @@
 if (empty($argv[1])) die("You need to pass e-mail address to this script");
 if (strpos($argv[1], '@') == false) die('You need valid e-mail address');
 $domain = trim( file_get_contents('/etc/hostname') );
+$lsbrelease = trim( shell_exec('/usr/bin/lsb_release -cs') );   #TODO Something more robut than this
 
 //Certbot changed completely how it functions
 //`apt-get remove certbot -y`;
@@ -14,12 +15,13 @@ $domain = trim( file_get_contents('/etc/hostname') );
 
 // 3rd time we have to change how certbot is installed *sigh*, they really like breaking old users, don't they? Since now via PIP, pray this works for more than a  week -Aleksi 22/08/2021
 
-if (!file_exists('/opt/certbot')) {
+#TODO this stuff should be on app installs ...
+if (!file_exists('/opt/certbot') && $lsbrelease == 'buster') {
     `python3 -m venv /opt/certbot/`;
     `/opt/certbot/bin/pip install --upgrade pip`;
     `/opt/certbot/bin/pip install certbot certbot-nginx`;
     `ln -s /opt/certbot/bin/certbot /usr/bin/certbot`;
-}
+} else echo `apt -y install certbot python3-certbot-nginx; `; #TODO Doesn't belong here for real ...
 
 if (!file_exists('/etc/letsencrypt/live/{$domain}')) echo `/usr/bin/certbot certonly -d {$domain} -n --nginx --agree-tos --email {$argv[1]}`;
